Fixed SSL typo
Based on the other examples, it looks like the ssl key file needs to be included as well as the cert. I added a command line option `--key` for that. Also, the SSL server was being given the cert file twice. This works for me now. Thanks for the great ws examples!