SimpleWebSocketServer: honor selectInterval=0
Currently for selectInterval=0 case, SimpleWebSocketServer removes the argument completely from the `select()` call, which means "no timeout".  For wrapping SimpleWebSocketServer in an async library, it's useful to support a timeout of 0 so that select returns immediately.

The cited behavior has been removed, so that selectInterval=0 and selectInterval=None are passed as-is to `select()`.  The default value of .1 seconds is unchanged.